Project_Starter_Files-Building_a_Metrics_Dashboard/reference-app/backend/requirements.txt
Flask==2.1.0
requests==2.23.0
pylint==2.4.0
pymongo==3.11.0
Flask-PyMongo==2.3.0
prometheus-flask-exporter==0.18.2
gunicorn==20.1.0

opentracing==2.4.0
jaeger-client==4.8.0
Flask-OpenTracing==1.1.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-exporter-jaeger